Rename `index.mjs` to `_index.mjs` in `@babel/runtime`
