+++ packages/babel-runtime-corejs2/helpers/esm/temporalRef.js
@@ -1,1 +1,1 @@
-export { default } from "../temporalRef/index.mjs"
+export { default } from "../temporalRef/_index.mjs"
