+++ packages/babel-runtime-corejs2/helpers/esm/toArray.js
@@ -1,1 +1,1 @@
-export { default } from "../toArray/index.mjs"
+export { default } from "../toArray/_index.mjs"
