+++ packages/babel-runtime/helpers/esm/toArray.js
@@ -1,1 +1,1 @@
-export { default } from "../toArray/index.mjs"
+export { default } from "../toArray/_index.mjs"
