+++ packages/babel-runtime/helpers/iterableToArray/_index.mjs
@@ -1,0 +1,3 @@
+export default function _iterableToArray(iter) {
+  if (typeof Symbol !== "undefined" && Symbol.iterator in Object(iter)) return Array.from(iter);
+}
